requirements.txt
bs4
django
django-browser-reload
django-compressor
django-cors-headers
django-unfold
djangorestframework
djangorestframework-simplejwt
lxml
pillow
requests
whitenoise
channels
daphne
channels-redis
python-dateutil
psycopg2
im
